Move convertPxToUnit to units module
