+++ lib/anime.esm.js
@@ -421,13 +421,6 @@ function getTransformUnit(propName) {
   if (stringContains(propName, 'rotate') || stringContains(propName, 'skew')) return 'deg';
 }
 
-// Values
-
-function getFunctionValue(val, animatable) {
-  if (!is.fnc(val)) return val;
-  return val(animatable.target, animatable.id, animatable.total);
-}
-
 function convertPxToUnit(el, value, unit) {
   const valueUnit = getUnit(value);
   if (arrayContains([unit, 'deg', 'rad', 'turn'], valueUnit)) return value;
@@ -435,7 +428,8 @@ function convertPxToUnit(el, value, unit) {
   if (!is.und(cached)) return cached;
   const baseline = 100;
   const tempEl = document.createElement(el.tagName);
-  const parentEl = (el.parentNode && (el.parentNode !== document)) ? el.parentNode : document.body;
+  const parentNode = el.parentNode;
+  const parentEl = (parentNode && (parentNode !== document)) ? parentNode : document.body;
   parentEl.appendChild(tempEl);
   tempEl.style.position = 'absolute';
   tempEl.style.width = baseline + unit;
@@ -444,6 +438,13 @@ function convertPxToUnit(el, value, unit) {
   const convertedUnit = factor * parseFloat(value);
   cache.CSS[value + unit] = convertedUnit;
   return convertedUnit;
+}
+
+// Values
+
+function getFunctionValue(val, animatable) {
+  if (!is.fnc(val)) return val;
+  return val(animatable.target, animatable.id, animatable.total);
 }
 
 function getCSSValue(el, prop, unit) {
