+++ lib/anime.umd.js
@@ -427,13 +427,6 @@
     if (stringContains(propName, 'rotate') || stringContains(propName, 'skew')) return 'deg';
   }
 
-  // Values
-
-  function getFunctionValue(val, animatable) {
-    if (!is.fnc(val)) return val;
-    return val(animatable.target, animatable.id, animatable.total);
-  }
-
   function convertPxToUnit(el, value, unit) {
     const valueUnit = getUnit(value);
     if (arrayContains([unit, 'deg', 'rad', 'turn'], valueUnit)) return value;
@@ -441,7 +434,8 @@
     if (!is.und(cached)) return cached;
     const baseline = 100;
     const tempEl = document.createElement(el.tagName);
-    const parentEl = (el.parentNode && (el.parentNode !== document)) ? el.parentNode : document.body;
+    const parentNode = el.parentNode;
+    const parentEl = (parentNode && (parentNode !== document)) ? parentNode : document.body;
     parentEl.appendChild(tempEl);
     tempEl.style.position = 'absolute';
     tempEl.style.width = baseline + unit;
@@ -450,6 +444,13 @@
     const convertedUnit = factor * parseFloat(value);
     cache.CSS[value + unit] = convertedUnit;
     return convertedUnit;
+  }
+
+  // Values
+
+  function getFunctionValue(val, animatable) {
+    if (!is.fnc(val)) return val;
+    return val(animatable.target, animatable.id, animatable.total);
   }
 
   function getCSSValue(el, prop, unit) {
